### 2021-08-02 #3555 @whedon
[PRE REVIEW]: CellPyLib: A Python Library for working with Cellular Automata
**Submitting author:** @lantunes (<a href="http://orcid.org/0000-0002-4867-5635">Luis Antunes</a>)
**Repository:** <a href="https://github.com/lantunes/cellpylib" target ="_blank">https://github.com/lantunes/cellpylib</a>
**Version:** v1.0.0
