### 2021-08-02 8e227fc
adding more docs and tests for Hopfield net
