### 2021-08-02 7ea45e8
adding docs on neighbourhood size
