### 2021-08-02 8ba7952
adding docs on the number of CA colors
