### 2021-08-02 11c09cb
adding docs for continuous CA
