+++ doc/totalistic.rst
@@ -1,5 +1,0 @@
-Totalistic CA
--------------
-
-#TODO
-https://github.com/lantunes/netomaton/blob/master/demos/totalistic_ca/README.md
