### 2021-08-02 82e3712
adding docs for Game of Life
