### 2021-08-02 d6b2d9b
adding docs for wireworld
