### 2021-08-02 d17cb4e
adding docs for Fredkin self-replicating CA
