### 2021-08-02 3784beb
adding docs for Hopfield net
