+++ paper/paper.bib
@@ -28,7 +28,8 @@
   number={1-3},
   pages={12--37},
   year={1990},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/0167-2789(90)90064-v}
 }
 
 @Article{harris2020array,
@@ -67,4 +68,21 @@
   publisher={IEEE COMPUTER SOC},
   doi={10.1109/MCSE.2007.55},
   year=2007
+}
+
+@article{nichele2017deep,
+  title={{Deep Learning with Cellular Automaton-based Reservoir Computing}},
+  author={Nichele, Stefano and Molund, Andreas},
+  year={2017},
+  publisher={Complex Systems Publications Inc}
+}
+
+@article{mordvintsev2020growing,
+  title={{Growing Neural Cellular Automata}},
+  author={Mordvintsev, Alexander and Randazzo, Ettore and Niklasson, Eyvind and Levin, Michael},
+  journal={Distill},
+  volume={5},
+  number={2},
+  pages={e23},
+  year={2020}
 }
