### 2021-08-02 71b213f
updating paper references
