+++ paper/paper.bib
@@ -74,7 +74,11 @@
   title={{Deep Learning with Cellular Automaton-based Reservoir Computing}},
   author={Nichele, Stefano and Molund, Andreas},
   year={2017},
-  publisher={Complex Systems Publications Inc}
+  volume={26},
+  number={4},
+  [ages={319--240},
+  publisher={Complex Systems Publications Inc},
+  doi={10.25088/complexsystems.26.4.319}
 }
 
 @article{mordvintsev2020growing,
