+++ paper/paper.bib
@@ -76,7 +76,7 @@
   year={2017},
   volume={26},
   number={4},
-  [ages={319--240},
+  pages={319--240},
   publisher={Complex Systems Publications Inc},
   doi={10.25088/complexsystems.26.4.319}
 }
