+++ paper/paper.bib
@@ -74,6 +74,7 @@
   title={{Deep Learning with Cellular Automaton-based Reservoir Computing}},
   author={Nichele, Stefano and Molund, Andreas},
   year={2017},
+  journal={Complex Systems},
   volume={26},
   number={4},
   pages={319--240},
