+++ paper/paper.bib
@@ -2,7 +2,8 @@
   title={{Cellular Automata: A Discrete Universe}},
   author={Ilachinski, Andrew},
   year={2001},
-  publisher={World Scientific Publishing Company}
+  publisher={World Scientific Publishing Company},
+  doi={10.1142/4702}
 }
 
 @book{wolfram2002new,
