### 2021-08-08 #3555 @lantunes
@jbytecode Thank you for agreeing to edit this submission! I have updated `paper.md` so that it includes `orcid` in the metadata section. Please let me know if there is any other metadata that needs to be added.
