### 2021-08-08 #3555 @blsqr
Hi everyone 👋

I'd be happy to review this submission. However, I cannot actively participate before 26th August – if that's ok, sign me up.

> assist in reviewing

What does *assist* entail in this context? Is there a distinction between reviewing and assisting?
