### 2021-08-08 #3555 @jbytecode
Dear @blsqr, the term `assist in reviewing` means `being a reviewer`. Thank you for accepting our invitation.

In JOSS, assigning adequate number of viewers may take time so we can wait for your availability.
