### 2021-08-12 #3555 @jbytecode
👋👋👋 @szhorvat  👋👋👋

Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software)?

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "CellPyLib: A Python Library for working with Cellular Automata". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/3555).
