### 2021-08-12 #3608 @whedon
[REVIEW]: CellPyLib: A Python Library for working with Cellular Automata
**Submitting author:** @lantunes (<a href="http://orcid.org/0000-0002-4867-5635">Luis Antunes</a>)
**Repository:** <a href="https://github.com/lantunes/cellpylib" target ="_blank">https://github.com/lantunes/cellpylib</a>
**Version:** v2.0.0
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.16977619" target="_blank">10.6084/m9.figshare.16977619</a>
