### 2021-08-12 #3555 @martibosch
Hello!

I see that there are already 2 reviewers assigned, but I'd also be happy to review this if you are up for taking 3 reviewers - this looks very interesting and it might actually be useful for my future research. I should note that I am quite familiar with land use/land cover CA but not so much about other applications - although I have read a few articles about their historical origins and first applications (e.g., artificial life).

Best,
Martí
