### 2021-08-12 #3608 @jbytecode
Dear reviewers @blsqr and @szhorvat

Thank you for accepting our invitation.

This is the reviewing page for the current submission. In this page, there are 20 checkboxes for each reviewer. These boxes indicate the corresponding reviewing tasks. Please fill them whenever you complete the corresponding task.

Please have a look at the [Reviewing Guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) page for details of reviewing process of JOSS.

Please do not hesitate to ask me anything.
