### 2021-08-31 lantunes/cellpylib#7 @blsqr
[JOSS] Minor tweaks to references in paper.md
Hi @lantunes, I'm here as one of the reviewers from your [submission to JOSS](https://github.com/openjournals/joss-reviews/issues/3608) 👋

I just read the paper (before looking into the code) and have only minor remarks regarding the references:

1) In the first paragraph, the two successive citations can be combined like this:

```
... Deep Learning [@nichele2017deep,@mordvintsev2020growing] ...
```

2) I think it would be more appropriate to mark up the title of the Wolfram book as italic (as done in the references) rather than verbatim. Just a suggestion, though.

3) For the Mordvintsev et al. citation, they supply the following BibTeX on their webpage, which includes a DOI (that @whedon must have missed):

```
@article{mordvintsev2020growing,
  author = {Mordvintsev, Alexander and Randazzo, Ettore and Niklasson, Eyvind and Levin, Michael},
  title = {Growing Neural Cellular Automata},
  journal = {Distill},
  year = {2020},
  note = {https://distill.pub/2020/growing-ca},
  doi = {10.23915/distill.00023}
}
```

4) Did you check if there are DOIs available for the Wolfram and von Neumann books? (There sometimes are, but if not, no worries.)

Let me know if there are questions or if I can assist with anything.
