### 2021-08-31 #3608 @blsqr
I just opened [an issue](https://github.com/lantunes/cellpylib/issues/7) regarding minor corrections to references and citation style.
