### 2021-09-19 #3608 @jbytecode
@lantunes would you consider adding a GitHub action that builds and tests the package after each commit so people can see the latest changes does not break the stuff?
