### 2021-09-19 lantunes/cellpylib#11 @szhorvat
[JOSS] Questions
This is not a bug report, just a means to ask questions @lantunes, for the purpose of the JOSS review. I hope it is okay to do it here. If not, can you please point me to a better location (perhaps there's a support forum)?
