### 2021-09-19 lantunes/cellpylib#11 @szhorvat
**Q1:** I am wondering if there is a simple way to use random sequential update order. I.e., select a random cell, update it, then another cell, update it, and so on. I know that a specific deterministic update order can be specified, but I am looking for a random one. Such an update would be used e.g. for an Ising model or the sandpile model.

**Q2:** Is it possible to have the update of a cell also affect its neighbours? Normally, the "updated" cell will change. There are models where it neighbours might change as well, for example the sandpile mode.

Generally, are you targeting stochastic cellular automata or only deterministic ones?
