### 2021-09-19 #3608 @szhorvat
I need a bit of help from  @lantunes  or reviewers who are more experienced with Python. Is it expected that animations (such as `plot2d_animate` from CellPyLib) won't work out of the box in Jupyter notebooks, when using inline graphics?
