### 2021-09-19 lantunes/cellpylib#11 @szhorvat
**Q3:** What is the `show_margins` argument of `plot2d_animate` for?  If using it in a Jupyter notebook, `show_margins=False` causes the figure to be output as a vector image whose width is the same as the browser window. If I use ipython in a terminal, it seems to result in an empty figure. It is not clear to me what the expected effect is.
