### 2021-09-19 lantunes/cellpylib#11 @szhorvat
**Q4:** I see no effects from the `scale` and `dpi` arguments of `plot2d_animate` in either a Jupyter notebook (tried `%matplotlib inline` and `%matplotlib notebook`) or ipython terminal (backend: MacOSX).  Can you please clarify?
