### 2021-09-19 #3608 @szhorvat
I opened a few issues in the CellPyLib repository, and I'd like to wait for responses from @lantunes before proceeding.

So far I enjoy playing with the library. There are nice tutorials which demonstrate classical concepts. The library is definitely useful to learning and teaching. Now I am trying to determine how suitable it is for research: exploring novel ideas that may not have been tried yet.

A general comment/suggestion to @lantunes (and to be clear: this is certainly not something to be fixed before publication in JOSS) is to think about which directions you want to go in in the future, and whether the current design of the library will generalize to incorporate new features. A few interesting things that might be added in the future:

 - Multi-dimensional CA. Currently, only 1D and 2D rectangular lattices are supported, each having their own dedicated functions. The current design does not seem to generalize to n-D.
 - Different lattice types, especially in 2D: triangular, hexagonal, or others.
 - A generalization of all of the above: CA on arbitrary networks (undirected or directed).
 - Support for stochastic CA, which usually require asynchronous update in random order. See the questions I asked in the issue tracker: https://github.com/lantunes/cellpylib/issues/11#issuecomment-922458207
 - Allow for a cell to change its neighbours as well when it is updated (as in the sandpile model).

If you consider all the different possibilities now, and decide which you might want to support in the future, it will help you come up with a future-proof design.
