### 2021-09-19 #3608 @szhorvat
>     * **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

This must be improved:

 1. There are basic steps for building the docs and running tests, but nothing explicit about whether contributions are welcome or how to contribute. This needs to be improved.
 2. This is implicit in the sense that the project is on GitHub, so people will assume they can use the issue tracker. IMO this is fine.
 3. There is no information about where to seek support. This forced me to [start asking questions on the issue tracker](https://github.com/lantunes/cellpylib/issues/11), which most other projects would not allow. This must be improved.
