### 2021-09-19 #3608 @szhorvat
>     * **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

https://cellpylib.org/ has a very clear and concise description of what the software can do _right now_. Great job on this! However, it would be nice to indicate what the _goals_ (as opposed to _current capabilities_) of the software project is, and _who the target audience is_. Is it trying to be a learning/teaching tool? I notice that it has built-in implementations of some classic models. Is it trying to be a research tool? Does it focus more on flexibility/generality or on performance (usually, there is a tradeoff between the two)? Does this project have the ambition to become the primary CA library for Python, which means that there should be a push to complete missing features?

While the current description is good, I feel that it can be improved by describing the future vision for the project.
