### 2021-09-22 #3608 @jbytecode
@lantunes could you please ping us (and if possible please update your status) if you receive the notifications from this issue?
