### 2021-09-22 #3608 @lantunes
Thank you for all of your comments and review thus far! I have been following the discussion, though haven't yet begun the process of addressing the individual items. I will begin addressing the items today, and aim to be done by the weekend.
