### 2021-09-22 #3608 @lantunes
> @lantunes would you consider adding a GitHub action that builds and tests the package after each commit so people can see the latest changes don't break the stuff?

@jbytecode Certainly! I will look into adding this.
