### 2021-09-22 b5b94f0
Adding action that runs tests
