+++ .github/workflows/python-app.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies and run tests with a single version of Python
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: CellPyLib
+name: Testing - main
 
 on:
   push:
