### 2021-09-22 c2b39a0
fixing typo in github action name
