### 2021-09-22 #3608 @lantunes
I have added a GitHub action to the cellpylib repo that runs the tests on any push to main. I have also added a status badge to the top of the project's README. @jbytecode
