+++ paper/paper.bib
@@ -84,11 +84,10 @@
 }
 
 @article{mordvintsev2020growing,
-  title={{Growing Neural Cellular Automata}},
-  author={Mordvintsev, Alexander and Randazzo, Ettore and Niklasson, Eyvind and Levin, Michael},
-  journal={Distill},
-  volume={5},
-  number={2},
-  pages={e23},
-  year={2020}
+  author = {Mordvintsev, Alexander and Randazzo, Ettore and Niklasson, Eyvind and Levin, Michael},
+  title = {Growing Neural Cellular Automata},
+  journal = {Distill},
+  year = {2020},
+  note = {https://distill.pub/2020/growing-ca},
+  doi = {10.23915/distill.00023}
 }
