### 2021-09-24 lantunes/cellpylib#7 @lantunes
Hi @blsqr,

Thank you for your review! I have pushed some changes directly to `master` which should address the items you mention. Regarding item 4, I was not able to find DOIs for the Wolfram and von Neumann books.
