### 2021-09-24 lantunes/cellpylib#7 @blsqr
Looking good! I just discovered that the DOI given for the Mordvintsev publication does not appear to work and filed an issue with them [here](https://github.com/distillpub/post--growing-ca/issues/13). Sorry that I didn't check that when proposing to use the BibTeX they provided.

It's a minor issue, so if they don't fix their DOI soon, I'd propose to revert to a citation of the webpage itself. Let's keep this issue open to not forget about it. Is that fine for you as well?
