### 2021-09-24 #3608 @blsqr
I opened a few additional issues in the [cellpylib project](https://github.com/lantunes/cellpylib/issues?q=is%3Aissue+is%3Aopen+JOSS) and would also pause my review at this point, continuing next week.

I believe I have arrived at a rather similar point as @szhorvat [above](https://github.com/openjournals/joss-reviews/issues/3608#issuecomment-922474364), where I am unclear about the scope of and long-term vision for the project. From what I've seen so far, cellpylib is well-suited for exploring CA-based complex systems and conveying principles of established examples from the scientific literature. Particularly the large number of tutorials and the well-written documentation are very inviting and enjoyable to read and follow along in code.

However, depending on the type of research that cellpylib aims to form the operational basis of, a number of additional features seem to become relevant; I would concur with the list given by @szhorvat, especially with function generalisations, different lattice types, and recursive rules.
Furthermore, performance aspects may become increasingly important.
This clearly can reach an unfeasibly large scope very easily…

@lantunes Can you perhaps comment on these points (here or in a separate issue) such that we can better align the review of the package to its overarching goals?
