### 2021-09-25 lantunes/cellpylib#7 @blsqr
Alright, the DOI issue was fixed in the distill repo, see https://github.com/distillpub/post--growing-ca/issues/13

I think we can close this issue now.
