### 2021-09-25 1850b3c
addressing test warnings
