+++ README.md
@@ -282,6 +282,11 @@ There are a number of unit tests for this project. To run the tests:
 $ python -m pytest tests
 ```
 
+If the `pytest-cov` package is installed, a coverage report can be generated by running the tests with:
+```
+$ python -m pytest tests/ --cov=cellpylib
+```
+
 --------------------
 
 ### Citation Info
