### 2021-09-28 8a950ba
adding test coverage for plotting functions
