### 2021-09-29 #3608 @lantunes
@szhorvat @blsqr Thank you for all your thoughts thus far. Regarding your comments on the future directions of the project, you certainly bring up very relevant and valid issues. The field of CA is vast, and there are indeed a great number of variations on the basic concept introduced decades ago, from stochastic behaviour to arbitrary underlying network structure, as you describe. I've thought about this in the past: how can one design a general-purpose CA library that reflects the full richness of the field? What I concluded is that there is a trade-off to be made in terms of straightforward and intuitive design and flexibility of model construction. In other words, if we want a library that can support the various kinds of CA that have been investigated in the literature, then the design of the library would necessarily be more complex, and possibly less intuitive to use. (There are likely other trade-offs as well, such as between flexibility and performance: as the library becomes more flexible in terms of the kinds of models one can construct, then it becomes more difficult to provide a performant implementation for any use case.)

To investigate this idea, I actually began creating another library some time ago, called Netomaton (https://github.com/lantunes/netomaton). Netomaton supports many more use cases than CellPyLib, including some of those in your list, @szhorvat. While some may find this library intuitive, I find that one needs to operate in the world of networks (this was the abstraction I chose; arguably, one can choose a better paradigm), and this may impose a burden on the user who is simply wishing to work with elementary CA. Moreover, I found that there are fewer opportunities to improve the performance with the underlying network abstraction.

In summary, I would say that the future scope of the project will not deviate too far from the kinds of systems that are currently supported. I am certainly open to supporting more model types and use cases, but not to the point that it requires a large redesign of the API. Elementary CA, and perhaps the other kinds of CA in the examples, are still very much a topic of research today, so even with some missing features, the library still has applicability in a research setting, in my opinion.
