### 2021-09-30 #3608 @blsqr
Thank you for the thorough response, @lantunes, and the pointer to Netomaton (which looks interesting as well!). I think the differences in abstractions between grid-based and network-based automata is a natural point at which to distinguish the libraries, with CellPyLib focussing on grid-based automata. As said, the trade-offs become substantial if aiming to do both…

I think the term "general-purpose" (also used in the software paper) is what may have let me astray, given the many different approaches to CA-based research and the corresponding trade-offs.
I would propose to extend the description of CellPyLib and the statement of need such that it is clearer that CellPyLib puts a focus on a subset of that field (elementary CA, CA with totalistic rules, …) and aims for a user-friendly interface and optimised and tested algorithms. In my eyes – given the vastness of the field – this would clearly be a feature of the library rather than a shortcoming, so I would highlight it accordingly. I think it's important for potential users to know what they can and cannot do with the library before spending time on getting to know and using it.
