### 2021-10-02 14ffcfe
adding docs and tests for bits_to_int and int_to_bits functions
