### 2021-10-02 44b7e8f
replacing copyright statement with link to Apache license
