### 2021-10-02 1321bc7
adding more documentation to functions
