### 2021-10-02 #3608 @lantunes
@jbytecode  @szhorvat @blsqr Thank you for all the suggestions thus far. They have been very constructive, and improve the library greatly. I am continuing to address the feedback in the `cellpylib` GitHub repo. I assumed I would have been done sooner, but other commitments have required more of my time. I hope to complete addressing all the feedback over the coming week.
