+++ cellpylib/__init__.py
@@ -1,9 +1,12 @@
-from .ca_functions import *
-from .rule_tables import *
-from .entropy import *
-from .ca_functions2d import *
-from .bien import *
-from .apen import *
-from .hopfield_net import *
-from .ctrbl_rule import *
-from .langtons_loop import *
+from .ca_functions import AsynchronousRule, ReversibleRule, binary_rule, init_simple, nks_rule, totalistic_rule, \
+    plot_multiple, bits_to_int, int_to_bits, init_random, plot, evolve
+from .rule_tables import random_rule_table, table_walk_through, table_rule
+from .entropy import mutual_information, average_cell_entropy, average_mutual_information, shannon_entropy, \
+    joint_shannon_entropy
+from .ca_functions2d import evolve2d, plot2d, plot2d_slice, plot2d_animate, plot2d_spacetime, init_simple2d, \
+    init_random2d, game_of_life_rule
+from .bien import binary_derivative, cyclic_binary_derivative, ktbien, tbien, bien
+from .apen import apen
+from .hopfield_net import HopfieldNet
+from .ctrbl_rule import CTRBLRule
+from .langtons_loop import LangtonsLoop
