### 2021-10-10 cb14dbe
adding BaseRule and using __call__
