### 2021-10-10 bf69027
adding x- and y-labels to 1D plot function #19
