### 2021-10-11 23096e3
linking to function and class definitions in docs #17
