+++ doc/working.rst
@@ -17,20 +17,21 @@ with CellPyLib:
                                     apply_rule=lambda n, c, t: cpl.nks_rule(n, 30))
 
 
-The initial conditions are instantiated using the function `init_simple`, which, in this example, creates a
-200-dimensional vector consisting of zeroes, except for the component in the center of the vector, which is initialized
-with a value of 1. Next, the system is subjected to evolution by calling the `evolve` function. The system evolves under
-the rule specified through the `apply_rule` parameter. Any function that accepts the three arguments `n`, `c` and `t`
-can be supplied as a rule, but in this case the built-in function `nks_rule` is invoked to provide Rule 30. The CA is
-evolved for 100 `timesteps`, or 100 applications of the rule to the initial and subsequent conditions.
+The initial conditions are instantiated using the function :py:func:`~cellpylib.ca_functions.init_simple`, which, in
+this example, creates a 200-dimensional vector consisting of zeroes, except for the component in the center of the
+vector, which is initialized with a value of 1. Next, the system is subjected to evolution by calling the
+:py:func:`~cellpylib.ca_functions.evolve` function. The system evolves under the rule specified through the
+``apply_rule`` parameter. Any function that accepts the three arguments ``n``, ``c`` and ``t`` can be supplied as a
+rule, but in this case the built-in function :py:func:`~cellpylib.ca_functions.nks_rule` is invoked to provide Rule 30.
+The CA is evolved for 100 ``timesteps``, or 100 applications of the rule to the initial and subsequent conditions.
 
-During each timestep, the function supplied to `apply_rule` is invoked for each cell. The `n` argument refers to the
+During each timestep, the function supplied to ``apply_rule`` is invoked for each cell. The ``n`` argument refers to the
 neighbourhood of the current cell, and consists of an array (in the 1-dimensional CA case) of the activities (i.e.
 states) of the cells comprising the current cell's neighbourhood (an array with length 3, in the case of a 1-dimensional
-CA with radius of 1). The `c` argument refers to index of the cell under consideration. It serves as a label identifying
-the current cell. The `t` argument is an integer specifying the current timestep.
+CA with radius of 1). The ``c`` argument refers to index of the cell under consideration. It serves as a label
+identifying the current cell. The ``t`` argument is an integer specifying the current timestep.
 
-Finally, to visualize the results, the `plot` function can be utilized:
+Finally, to visualize the results, the :py:func:`~cellpylib.ca_functions.plot` function can be utilized:
 
 .. code-block::
 
@@ -61,8 +62,9 @@ Initializing CA
 
 A CA is initialized by simply instantiating an array of an array (for 1-dimensional CA), or an array of an array of an
 array (for 2-dimensional CA). This will represent the initial conditions of the system, which can be submitted to the
-`evolve` function. For convenience, there are several built-in functions for common CA initializations. For example, the
-`init_simple` function can be used to create a 1D binary initialization with a 1 in the center:
+:py:func:`~cellpylib.ca_functions.evolve` function. For convenience, there are several built-in functions for common CA
+initializations. For example, the :py:func:`~cellpylib.ca_functions.init_simple` function can be used to create a 1D
+binary initialization with a 1 in the center:
 
 .. code-block::
 
@@ -71,41 +73,46 @@ array (for 2-dimensional CA). This will represent the initial conditions of the 
     print(cellular_automaton)
     # [[0 0 0 0 0 1 0 0 0 0]]
 
-An analogous function, `init_simple2d`, exists for 2-dimensional CA.
[…]

[…]
