### 2021-10-11 lantunes/cellpylib#11 @lantunes
> This is not a bug report, just a means to ask questions @lantunes, for the purpose of the JOSS review. I hope it is okay to do it here. If not, can you please point me to a better location (perhaps there's a support forum)?

@szhorvat Thank you very much for the questions! Unfortunately, there isn't a dedicated forum for this library at the moment. If usage increases, I will create a Google group, but for now, creating an issue here is the best way.
