### 2021-10-11 cab02b7
adding Support section to README
