+++ README.md
@@ -286,6 +286,13 @@ If the `pytest-cov` package is installed, a coverage report can be generated by 
 ```
 $ python -m pytest tests/ --cov=cellpylib
 ```
+--------------------
+
+### Support
+
+If you have any questions or comments, or if you find any bugs, please open an issue in this project. Please feel free
+to fork the project, and create a pull request, if you have any improvements or bug fixes. We welcome all feedback and
+contributions.
 
 --------------------
 
