### 2021-10-11 d34285a
adding example for updating cell randomly; fixing 2D support in AsynchronousRule #11
