### 2021-10-16 3f1861a
adding imshow keyword args to plotting function signatures
