### 2021-10-16 4562162
adding option that controls whether or not plots are shown
