### 2021-10-16 067fae4
adding show_grid and other arguments to plot2d functions
