+++ cellpylib/ca_functions2d.py
@@ -5,7 +5,8 @@ import matplotlib.collections as mcoll
 import numpy as np
 
 
-def plot2d(ca, timestep=None, title='', *, colormap='Greys', show=True, **imshow_kwargs):
+def plot2d(ca, timestep=None, title='', *, colormap='Greys', show_grid=False, show_margin=True, scale=0.6,
+           show=True, **imshow_kwargs):
     """
     Plots the state of the given 2D cellular automaton at the given timestep. If no timestep is provided, then the last
     timestep is plotted.
@@ -18,22 +19,38 @@ def plot2d(ca, timestep=None, title='', *, colormap='Greys', show=True, **imshow
 
     :param colormap: the color map to use (default is "Greys")
 
+    :param show_grid: whether to display a grid (default is False)
+
+    :param show_margin: whether to display the margin (default is True)
+
+    :param scale: the scale of the figure (default is 0.6)
+
     :param show: show the plot (default is True)
 
     :param imshow_kwargs: keyword arguments for the Matplotlib `imshow` function
     """
     cmap = plt.get_cmap(colormap)
+    fig, ax = plt.subplots()
     plt.title(title)
+    if not show_margin:
+        fig.subplots_adjust(left=0, bottom=0, right=1, top=1, wspace=0, hspace=0)
     if timestep is not None:
         data = ca[timestep]
     else:
         data = ca[-1]
-    plt.imshow(data, interpolation='none', cmap=cmap, **imshow_kwargs)
+
+    _add_grid_lines(ca, ax, show_grid)
+
+    im = plt.imshow(data, interpolation='none', cmap=cmap, **imshow_kwargs)
+    if not show_margin:
+        baseheight, basewidth = im.get_size()
+        fig.set_size_inches(basewidth*scale, baseheight*scale, forward=True)
     if show:
         plt.show()
 
 
-def plot2d_slice(ca, slice=None, title='', *, colormap='Greys', show=True, **imshow_kwargs):
+def plot2d_slice(ca, slice=None, title='', *, colormap='Greys', show_grid=False, show_margin=True, scale=0.6,
+                 show=True, **imshow_kwargs):
     """
     Plots a slice through the evolved states of a 2D cellular automaton. For example, consider the following `ca`,
     which may represent the evolution of a 3x3 2D cellular automaton over 3 timesteps:
@@ -74,17 +91,32 @@ def plot2d_slice(ca, slice=None, title='', *, colormap='Greys', show=True, **ims
 
     :param colormap: the color map to use (default is "Greys")
 
+    :param show_grid: whether to display a grid (default is False)
+
+    :param show_margin: whether to display the margin (default is True)
+
+    :param scale: the scale of the figure (default is 0.6)
+
     :param show: show the plot (default is True)
 
     :param imshow_kwargs: keyword arguments for the Matplotlib `imshow` function
     """
     cmap = plt.get_cmap(colormap)
+    fig, ax = plt.subplots()
     plt.title(title)
+    if not show_margin:
+        fig.subplots_adjust(left=0, bottom=0, right=1, top=1, wspace=0, hspace=0)
     if slice is not None:
         data = ca[:, slice]
     else:
         data = ca[:, len(ca[0])//2]
-    plt.imshow(data, interpolation='none', cmap=cmap, **imshow_kwargs)
+
+    _add_grid_lines(ca, ax, show_grid)
+
+    im = plt.imshow(data, interpolation='none', cmap=cmap, **imshow_kwargs)
[…]
