### 2021-10-16 lantunes/cellpylib#11 @lantunes
With respect to **Q3**, the `show_margins` argument controls whether or not a margin is displayed in the resulting plot. This may be desirable when creating animations that are meant to be saved and published. The Wireworld demos are currently the only places where this argument is used. When `show_margins` is set to `False`, then the plot takes up the entirety of the window. The `scale` argument, which is only used when `show_margins` is `False`, determines the resulting size of the image.
