### 2021-10-16 f91e83f
adding clarifying comments to plot2d functions
