### 2021-10-16 lantunes/cellpylib#11 @lantunes
With respect to **Q4**, the `scale` argument is only used when the `show_margin` argument is `False`. It controls the resulting scale (i.e. relative size) of the image when there are no margins. The `dpi` argument represents the dots per inch of the animation when it is saved. There will be no visible effect of the `dpi` argument if the animation is not saved (i.e. when `save` is `False`).

I have added more information to the "plot2d" function docstrings to clarify. The change is on the `joss_review` branch. The relevant commit is f91e83fd34ef716a5f5808031535a97c9a4f1133.
