### 2021-10-16 64d3927
checking for warning message in plot tests
