### 2021-10-16 37ec83c
not asserting warnings in plot tests because in some environments they are not emitted
