### 2021-10-16 3362079
adding clarifying comment about library scope to README and paper
