### 2021-10-16 #3608 @lantunes
@blsqr regarding you comment about the scope of the library, I have added a sentence to the `README.md`, and a *Scope* section to the paper, to clarify this issue. The relevant commit is here: https://github.com/lantunes/cellpylib/pull/23/commits/336207968dfe8d9cd46120d10f9cd6184e4b45e3.
