### 2021-10-18 lantunes/cellpylib#24 @blsqr
[JOSS] Highlight target audience in documentation
As a last issue for the JOSS review (at least from my side), there is the following point to address in the documentation:

>  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

The current version of the documentation is quite clear in describing what the software is designed to do. However, I think it could be improved by more clearly stating the target audience. While the information is there indirectly (people interested in working with the mentioned CAs), I'd like to suggest to add a paragraph that *directly* addresses the target audience and answers the question of *why* they should use `CellPyLib` (and in which cases an alternative may be more appropriate).

I think that a condensed version of the section you use in the software paper could be very helpful to have on the landing page of the documentation. Maybe a short list of bullet points (e.g. *"You should use `CellPyLib` if..."*) would also suffice.
