### 2021-10-18 #3608 @blsqr
Thanks, @lantunes! I have left a few minor comments in https://github.com/lantunes/cellpylib/pull/23#pullrequestreview-782006416 and have opened an issue (https://github.com/lantunes/cellpylib/issues/24) regarding my last open point in the review checklist.
