### 2021-10-19 03d9d35
using TypeError or ValueError in place of Exception
