### 2021-10-24 b9b2a6f
keeping version in module init file instead of setup script
