### 2021-10-30 #3608 @szhorvat
@lantunes @blsqr Does the test suite pass for you when using the current `joss_review` branch?

I get:

```
AttributeError: 'PathCollection' object has no property 'depthshade'
```

It is coming from:

https://github.com/lantunes/cellpylib/blob/joss_review/cellpylib/ca_functions2d.py#L158
