### 2021-10-30 lantunes/cellpylib#11 @szhorvat
> With respect to **Q2**, my understanding is that a Sandpile is a regular 2D cellular automaton. That is, the update rule is applied synchronously to all cells. Thus, the neighbours of a cell are affected by an update to the cell, but I don't believe there is any kind of asynchronous update scheme in the Sandpile model. I have added a built-in Sandpile rule on the `joss_review` branch. (The relevant commit is [045a209](https://github.com/lantunes/cellpylib/commit/045a209ddd86376ab23eaad30dc568bdca210d46).) Hopefully it will aid in this discussion. In any case, I suspect that the kind of model you may be describing could be implemented with an `AsynchronousRule`.

Yes, you are right, we can think of it as synchronous, though perhaps this is not the most efficient way to simulate it.

I am wondering what the best way would be to do the following:

 1. Add a grain of sand at the middle (alternative: at a random location)
 2. Simulate the system until it reaches a fixed point (nothing changes anymore). Is there a way to detect a fixed point and automatically stop the simulation?
 3. Repeat from step 1.

What would you recommend to do this with a reasonable performance?
