### 2021-10-31 #3608 @lantunes
@szhorvat The test suite is passing for me locally on the `joss_review` branch. The test suite is also passing when run during the Github action workflow for the project. What version of Matplotlib are you using in your environment?
