### 2021-10-31 0c375d5
using Axes3D when plotting 2D spacetime
