### 2021-10-31 #3608 @lantunes
@szhorvat I managed to reproduce this test failure with the following environment:
```
- python=3.7
- matplotlib=3.4.2
- numpy=1.16
- pytest
- pytest-cov
```
I pushed a commit to the `joss_review` branch that should fix this issue.
