### 2021-10-31 8ef7ca4
matrix testing python 3.6 and 3.7 with associated dependencies
