+++ .github/workflows/python-app.yml
@@ -13,18 +13,29 @@ jobs:
   build:
 
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        include:
+          - python-version: 3.6
+            numpy-version: 1.15.4
+            matplotlib-version: 3.0.2
+          - python-version: 3.7
+            numpy-version: 1.16.6
+            matplotlib-version: 3.4.3
 
     steps:
     - uses: actions/checkout@v2
-    - name: Set up Python 3.6
+    - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v2
       with:
-        python-version: 3.6
+        python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         pip install pytest
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        pip install numpy==${{ matrix.numpy-version }}
+        pip install matplotlib==${{ matrix.matplotlib-version }}
     - name: Test with pytest
       run: |
         pytest
+
