### 2021-10-31 lantunes/cellpylib#11 @lantunes
@szhorvat I have made some changes to the library to support what you describe with sandpiles. The relevant commit is 70aa94e4089c7beb0c977177daf83eb52b7de995.

In summary, I have made the `timesteps` parameter of the `evolve` and `evolve2d` functions be either an int or a callable. If one knows the number of timesteps in advance, an int can be supplied. If one doesn't know the number of timesteps in advance, then a callable can be supplied. This callable accepts the full history of the CA evolution up to the point when it is called, along with the number of the current timestep. The callable must return `False` if evolution is to halt, and `True` otherwise.

This change allows one to detect when a fixed point has been reached. Also, adding a grain of sand to a particular cell at a particular timestep should be trivial. I have added convenience functions to the `Sandpile` class to perform these operations. See the `Sandpile.until_fixed_point()` and `Sandpile.add_grain()` functions. I have also updated the documentation for the Sandpile tutorial to illustrate these capabilities.
