### 2021-10-31 lantunes/cellpylib#11 @szhorvat
> See the `Sandpile.until_fixed_point()`

I am wondering why this was made part of the `Sandpile` class, since it's a generally useful function, not tied in any way to sandpiles.

Also, to be clear, I am not expecting you to implement everything I asked about as part of the library. That would of course be unreasonable. I am trying to judge how usable the library is for arbitrary CA-related tasks, some of which you may not have anticipated. The sandpile model and the Ising model were two such CAs.
