### 2021-10-31 lantunes/cellpylib#11 @szhorvat
> Also, adding a grain of sand to a particular cell at a particular timestep should be trivial.

My concern is: Is there a performance difference between calling `evolve2d` with 1 timestep `n` times vs calling `evolve2d` with `n` timesteps once?  If not, then adding grains is indeed trivial.

Generally, the current sandpile implementation seems to be rather slow.  I am wondering how one would go about generating an image like this one (but of course smaller): https://en.wikipedia.org/wiki/Abelian_sandpile_model#/media/File:Sandpile_on_infinite_grid,_3e7_grains.png

Here's my try:

```
n = 50
sandpile = cpl.Sandpile(n, n)
ca = cpl.init_simple2d(n, n, val=5)

for i in range(100):
    ca=np.array([ca[-1]])
    ca[0,n//2,n//2]=5
    ca=cpl.evolve2d(ca, apply_rule=sandpile, timesteps=cpl.Sandpile.until_fixed_point(), neighbourhood='von Neumann')

cpl.plot2d(ca)
```

This is unfortunately incredibly slow, taking multiple minutes, despite generating a tiny result.

Would this code be how you envisioned people to use the library, or do you have recommendations for changes?
