### 2021-10-31 lantunes/cellpylib#11 @szhorvat
This brings me to **Q5:**

What is the recommended way to construct an continuous series of CA states if calling `evolve` / `evolve2d` multiple times?  Suppose I create an initial configuration `ca`, which will be an 1-by-n array (for a 1D CA). I let it evolve for `k` time steps, getting a `k+1`-by-n result. Now I modify the last state manually, and I would like to continue the evolution. If I call `evolve` again, it will take `ca[0]` as the starting state and discard everything else.

What is the recommended way to continue the evolution from the last recorded state, and retain all previously computed states (e.g. for an animation in 2D)?
