### 2021-11-01 lantunes/cellpylib#11 @lantunes
@szhorvat In terms of the performance, given the scenario you describe, I expect it would be slower to evolve the CA for 1 timestep `n` times, since there is initialization and allocation of an array each time, but I'm not sure how much slower. It all depends on what you're trying to do. The `Sandpile` class provides support for adding a grain at particular locations and timesteps over the course of a single evolution, but that assumes you know on what timesteps you'd like to do the addition. If it's not obvious what those timesteps are, then yes, you would have to evolve the CA until it meets some condition, and then restart the evolution after adding the grain. I hope I am following correctly, and this answers your question.

It's true that the library in general is not optimized for speed. For your particular example, I tried running it on my machine, and it required 159 seconds. I tried using a `Moore` neighbourhood instead of a `von Neumann` neighbourhood, and it completed in 32 seconds. The reason is that I use a Numpy masked array when the neighbourhood is `von Neumann`, and this makes things considerably slower. However, 32 seconds is still not ideal. The problem is that there is a trade-off to be made between library generality (i.e. offering flexibility in terms of what kinds of models may be implemented) and performance (the amount of time required for the system to evolve). A CA may be given a very fast implementation if we know the rule in advance, and we can optimize the operations for the rule in question. Even so, there are ways that one may make CellPyLib faster, such as by incorporating parallelization. (Though a complicating factor is that a rule can be any arbitrary callable.) But this is out of scope for the current release, and it will be the focus of development for future releases.

Regarding **Q5**, I have provided an example of what I believe you describe:
```
import cellpylib as cpl
import numpy as np

n = 50
sandpile = cpl.Sandpile(n, n)
history = [cpl.init_simple2d(n, n, val=5)[-1]]

for i in range(100):
    initial = np.array([history[-1]])
    initial[0, n//2, n//2] = 5
    ca = cpl.evolve2d(initial, apply_rule=sandpile, timesteps=cpl.Sandpile.until_fixed_point(), neighbourhood='Moore')
    history.extend(ca)

cpl.plot2d_animate(np.array(history))
```
Basically, I keep a separate list, `history`, that I will extend after each call to `evolve2d`. Is this an accurate interpretation of the process you describe?
