### 2021-11-01 a3c23e0
moving until_fixed_point function out of Sandpile class
