+++ demos/sandpile_add_grain_demo.py
@@ -10,7 +10,7 @@ sandpile.add_grain(cell_index=(23, 23), timestep=1)
 initial = np.loadtxt('sandpile_add_grain_demo.txt', dtype=int)
 initial = np.array([initial])
 
-ca = cpl.evolve2d(initial, timesteps=sandpile.until_fixed_point(),
+ca = cpl.evolve2d(initial, timesteps=cpl.until_fixed_point(),
                   apply_rule=sandpile, neighbourhood="von Neumann")
 
 print("Number of timesteps to reach fixed point: %s" % len(ca))
