### 2021-11-01 lantunes/cellpylib#11 @lantunes
>I am wondering why this was made part of the Sandpile class, since it's a generally useful function, not tied in any way to sandpiles.

Good point! I have moved the `until_fixed_point` function out of the `Sandpile` class. The relevant commit is a3c23e04405f8ab38184c2d57fc630c0417c7658.
