### 2021-11-02 f774faa
improving language in bits_to_int docstring
