### 2021-11-02 9db36d9
adding description of target audience to paper and README
