### 2021-11-02 lantunes/cellpylib#24 @lantunes
@blsqr I have updated the paper and README files in an attempt to describe the target audience. The relevant commit is 9db36d9518d385310d5919f742f07814a1faebcc.
