### 2021-11-02 lantunes/cellpylib#24 @blsqr
@lantunes Great changes! I think this really helps to decide whether CellPyLib is the right tool for the job 👍
