### 2021-11-08 #3608 @lantunes
@jbytecode Yes, I believe I have. All the changes are on the `joss_review` branch, in this PR: https://github.com/lantunes/cellpylib/pull/23. The changes are ready to be merged.
