### 2021-11-08 lantunes/cellpylib#11 @szhorvat
Re Q5 above:

I wonder why not just let `evolve2d` continue from the _last_ timestep of its input array instead of from the first, and return both the prior timesteps provided as input, as well as the newly computed timesteps? What do you think?
