### 2021-11-08 #3608 @szhorvat
@jbytecode I'm done checking off the items.

I think this is a nice library, though it's still early days, and more battle testing will certainly reveal opportunities for improvements. The quality has substantially improved during the review. Performance may end up being the first limiting factor for some users, though that is the nature of Python (trading off ease of use for performance).

I recommend publication in JOSS.
