### 2021-11-10 #3608 @blsqr
@jbytecode I am also done with the checklist but forgot to write: *I endorse publication of the CellPyLib package and paper in JOSS*. 🚀
I think this is a very valuable package, especially because of the many well-written tutorials and its ease of use for exploring and studying the behaviour of CA.

Thanks for letting me be part of this review!
