### 2021-11-10 #3608 @jbytecode
Dear @martibosch

Making a final decision with the reports at hand is possible because we have two positive ones.

But, of course, if you want to finalize your review until November, 13th (Saturday), we can wait a little bit.

Please let me know your what you decide to do.

Thank you.
