### 2021-11-10 #3608 @jbytecode
Dear @lantunes

I have just read the final form of the paper and it seems it is in a very good state. However, in the code snipped, the code flows the page width.

is it possible to define lambda before the function call and pass it to function as an argument so we can gain some extra space?
