+++ paper/paper.md
@@ -91,8 +91,9 @@ import cellpylib as cpl
 
 cellular_automaton = cpl.init_simple(200)
 
+rule = lambda n, c, t: cpl.nks_rule(n, 30)
 cellular_automaton = cpl.evolve(cellular_automaton, timesteps=100, 
-                                apply_rule=lambda n, c, t: cpl.nks_rule(n, 30))
+                                apply_rule=rule)
 ```
 
 First, the initial conditions are instantiated using the function `init_simple`, which, in this example, creates a 
