### 2021-11-10 #3608 @lantunes
@jbytecode I have updated the paper on the `joss_review` branch, so that the lambda is defined before the function call. I believe this addresses the issue.
