### 2021-11-10 #3608 @jbytecode
Dear @lantunes

- [ ] In final draft of the paper, (in the code snippet) an extra space between the lambda definition and the function call would be good but it is  not really necessary.

- [ ] Please re-check the whole paper by means of both language and format (one more time). When it is finished, please ping me here.

- [ ] make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)) as described [here](<https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (**title** and **author list**), or edit these if that is not the case.

- [ ] use the format vx.y.z as your versioning format.

Thank you in advance.
