+++ paper/paper.md
@@ -92,6 +92,7 @@ import cellpylib as cpl
 cellular_automaton = cpl.init_simple(200)
 
 rule = lambda n, c, t: cpl.nks_rule(n, 30)
+
 cellular_automaton = cpl.evolve(cellular_automaton, timesteps=100, 
                                 apply_rule=rule)
 ```
