### 2021-11-10 #3608 @lantunes
@jbytecode I have added an extra space between the lambda definition and the function call, and I have re-checked the whole paper. Should I go ahead and make a tagged release and archive?
