### 2021-11-10 #3608 @lantunes
@jbytecode I have created a tagged release and deposited the files in figshare. The version number is `v2.0.0` and the DOI is `10.6084/m9.figshare.16977619`.
