### 2021-11-10 #3608 @szhorvat
It's not critical for the review (which is why I went ahead and concluded it here), but if you're going to make a major new release, it might make sense to think about minor breaking changes to the API that can improve the user experience. What do you think about my suggestion here? https://github.com/lantunes/cellpylib/issues/11#issuecomment-963535331 Wouldn't it make it so much easier to continue a simulation, which I expect will be something that many people want to do, _especially_ when having the new `until_fixed_point()`?
