### 2021-11-10 #3608 @szhorvat
I might have missed something, but what there a breaking change so far which warrants a 2.0.0 according to semantic versioning?
