### 2021-11-10 #3608 @jbytecode
- @szhorvat - I am agreed with the versioning scheme.
- @lantunes what do you think for this?
- @szhorvat - should I break the editorial stuff and wait for your discussion now?
