### 2021-11-10 #3608 @szhorvat
>     * @szhorvat - should I break the editorial stuff and wait for your discussion now?

I'm sorry for the misunderstanding, please do not. This is a nice package, and I am happy to see it published in JOSS. My review is of course final.

The above message was a reaction to seeing the 2.0.0 version number, which surprised me (is it necessary to jump to 2.0? if doing that jump, perhaps take the opportunity to break API?) It is of course up to @lantunes what direction he takes his package, and what gets included in any given release. It would have been better to take my suggestion to the cellpylib repo instead of the review thread.
