### 2021-11-10 #3608 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
