### 2021-11-11 #3608 @lantunes
@szhorvat I apologize for not responding sooner to your last comment in lantunes/cellpylib#11. I agree with your suggestion regarding adding the ability to specify initial conditions with a prior history. I have created an issue, lantunes/cellpylib#25, to make this change. It will be released in `v2.1.0`, which hopefully will be released soon.

@szhorvat @jbytecode Regarding the decision to issue a new major release (i.e. `v2.0.0`), there were certain changes to the library, such as specifying explicitly which symbols are imported in the `__init__.py` file, and making built-in rules like `AsynchronousRule` callable (and removing the `apply_rule` function), which could result in breaking changes for users on `v1`. There were also sufficient changes to the API, such as the interpretation of the `timesteps` parameter, the enhancements to the plotting functions, and the additon of new built-in models, like `Sandpile`, that I felt a major release was warranted.

@kyleniemeyer Thank you!

@jbytecode Thank you for your suggestions, and for organizing the review process!

@szhorvat @blsqr I would like to especially thank both of you for all of the time and effort you put into reviewing my submission. Your comments and suggestions were extremely helpful, and definitely made the library much better. I am very grateful for your interest in the library, and for your insights and feedback.
